fix: make page.title != page.slug comparison case insensitive
I don't exactly understand why this was necessary; the behavior I
observed was that for a result node that is untitled, it would receive a
title like "000b" by default, whereas its slug is "000B". This resulted
in such a node having a spurious title displayed.

I'm not sure if what I did here is the best way to fix this, so feel
free to change it.